@@ -3,12 +3,13 @@
   "description": "cypress testing",
   "version": "1.0.0",
   "dependencies": {
-    "@xapi/cmi5": "^1.1.1",
-    "@xapi/xapi": "^1.2.0",
-    "cypress": "^8.7.0",
+    "@xapi/cmi5": "^1.1.4",
+    "@xapi/xapi": "^2.0.1",
+    "cypress": "^12.13.0",
     "cypress-fill-command": "^1.0.2",
     "cypress-image-snapshot": "^4.0.1",
-    "typescript": "^4.3.2"
+    "typescript": "^5.1.3",
+    "wait-on": "^7.0.1"
   },
   "resolutions": {
     "axios": "0.21.1"
@@ -21,6 +22,6 @@
     "cy:verify": "npx cypress verify"
   },
   "devDependencies": {
-    "npm-check-updates": "^16.10.7"
+    "@simonsmith/cypress-image-snapshot": "^7.0.0"
   }
 }
