@@ -5,197 +5,112 @@ Permission to use, copy, modify, and distribute this software and its documentat
 The full terms of this copyright and license should always be found in the root directory of this software deliverable as "license.txt" and if these terms are not found with this software, please contact the USC Stevens Center for the full license.
 */
 import React, { useState } from "react";
-import { makeStyles } from "tss-react/mui";
-import { Grid } from "@mui/material";
-import { Typography } from "@mui/material";
+import { Grid, Typography, useMediaQuery, Paper, styled } from "@mui/material";
 import { Target } from "types";
 import clsx from "clsx";
 import withLocation from "wrap-with-location";
+import { isNoHeader, shouldDisplayPortrait } from "utils";
+import {
+  surveySaysStylesDesktop,
+  surveySaysStylesMobile,
+} from "styles/surveySays";
+
+const Item = styled(Paper)(({ theme }) => ({
+  backgroundColor: "#1B6A9C !important",
+  ...theme.typography.body2,
+  textAlign: "center",
+  color: "#fff",
+  borderRadius: 10,
+}));
 
 const SurveySays = (props: {
   search: { lesson: string };
   hasMedia: boolean;
   targets: Target[];
+  isMobile: boolean;
 }): JSX.Element => {
-  const { classes: styles } = useStyles();
+  const { classes: stylesDesktop } = surveySaysStylesDesktop();
+  const { classes: stylesMobile } = surveySaysStylesMobile();
   const [expandedCard, setExpandedCard] = useState(-1);
+  const matchesMobile = useMediaQuery("(max-width : 600px)", { noSsr: true });
+
+  const surveySaysGrid = (
+    <Grid
+      container
+      rowSpacing={shouldDisplayPortrait() || matchesMobile ? 1 : 3}
+      columnSpacing={{ xs: 1, sm: 2, md: 3 }}
+    >
+      {props.targets.map((target, idx) => (
+        <Grid item xs={12} lg={6} key={idx} columnSpacing={4} rowSpacing={3}>
+          <Item>
+            <div
+              id={`card-${idx}`}
+              className={clsx({
+                [stylesDesktop.fixedSurveyCard]: idx !== expandedCard,
+                [stylesMobile.fixedSurveyCardMobile]:
+                  (shouldDisplayPortrait() || matchesMobile) &&
+                  idx !== expandedCard,
+                [stylesDesktop.expandableSurveyCard]: idx === expandedCard,
+                [stylesDesktop.completeSatisfied]:
+                  target.status === "complete" && target.score === 1,
+                [stylesDesktop.completeUnsatisfied]:
+                  target.status === "complete" && target.score !== 1,
+                [stylesDesktop.default]: target.status !== "complete",
+              })}
+              onClick={() => {
+                if (idx === expandedCard) {
+                  setExpandedCard(-1);
+                } else {
+                  setExpandedCard(idx);
+                }
+              }}
+            >
+              {idx === expandedCard && target.status === "complete" ? (
+                <Typography
+                  variant={target.status !== "complete" ? "h6" : "caption"}
+                >
+                  {target.status === "complete" ? target.text : idx + 1}
+                </Typography>
+              ) : (
+                <Typography
+                  className={clsx({
+                    [stylesDesktop.fixedSurveyCardText]: true,
+                    [stylesDesktop.centerLock]: true,
+                  })}
+                  variant={target.status !== "complete" ? "h6" : "caption"}
+                  style={{
+                    paddingLeft: 10,
+                    paddingRight: 10,
+                    width: "100%",
+                    boxSizing: "border-box",
+                  }}
+                >
+                  {target.status === "complete" ? target.text : idx + 1}
+                </Typography>
+              )}
+            </div>
+          </Item>
+        </Grid>
+      ))}
+    </Grid>
+  );
 
   return (
     <>
       <div
         className={clsx({
-          [styles.bodyRoot]: true,
-          [styles.bodyNoMedia]: !props.hasMedia,
-          [styles.bodyMedia]: props.hasMedia,
+          [stylesDesktop.bodyRoot]: true,
+          [stylesDesktop.bodyRootNoHeader]: isNoHeader(),
+          [stylesDesktop.bodyNoMedia]: !props.hasMedia,
+          [stylesDesktop.bodyMedia]: props.hasMedia,
         })}
       >
-        <div className={styles.surveyInnerBoard}>
-          <div className={styles.survey}>
-            <Grid container>
-              <Grid item xs={12}>
-                <Typography
-                  className={styles.surveyQuestion}
-                  variant="h6"
-                  style={{ width: "100%" }}
-                >
-                  Key Takeaways
-                </Typography>
-              </Grid>
-              {props.targets.map((target, idx) => {
-                return (
-                  <Grid item xs={12} key={idx} style={{ margin: 8 }}>
-                    <div
-                      id={`card-${idx}`}
-                      className={clsx({
-                        [styles.fixedSurveyCard]: idx !== expandedCard,
-                        [styles.expandableSurveyCard]: idx === expandedCard,
-                        [styles.completeSatisfied]:
-                          target.status === "complete" && target.score === 1,
-                        [styles.completeUnsatisfied]:
-                          target.status === "complete" && target.score !== 1,
-                        [styles.default]: target.status !== "complete",
-                      })}
-                      onClick={() => {
-                        if (idx === expandedCard) {
-                          setExpandedCard(-1);
-                        } else {
-                          setExpandedCard(idx);
-                        }
-                      }}
-                    >
-                      {idx === expandedCard && target.status === "complete" ? (
-                        <Typography
-                          variant={
-                            target.status !== "complete" ? "h6" : "caption"
-                          }
-                        >
-                          {target.status === "complete" ? target.text : idx + 1}
-                        </Typography>
-                      ) : (
-                        <Typography
-                          className={clsx({
-                            [styles.fixedSurveyCardText]: true,
-                            [styles.centerLock]: true,
-                          })}
-                          variant={
-                            target.status !== "complete" ? "h6" : "caption"
-                          }
-                          style={{
-                            paddingLeft: 10,
-                            paddingRight: 10,
-                            width: "100%",
-                            boxSizing: "border-box",
-                          }}
-                        >
-                          {target.status === "complete" ? target.text : idx + 1}
-                        </Typography>
-                      )}
-                    </div>
-                    {/* <div
-                      className={clsx({
-                        [styles.expandableSurveyCard]: true,
-                        [styles.completeSatisfied]:
-                          target.status === "complete" && target.score === 1,
-                        [styles.completeUnsatisfied]:
-                          target.status === "complete" && target.score !== 1,
-                        [styles.default]: target.status !== "complete",
-                      })}
-                    >
-                      <Typography variant={
-                          target.status !== "complete" ? "h6" : "caption"
-                        }>
-                      Some super super super super super super super super super super super super long text that really need a massive overflow to show these really cool expectaions that you somehow got!
-                      </Typography>
-                    </div> */}
-                  </Grid>
-                );
-              })}
-            </Grid>
-          </div>
+        <div className={stylesDesktop.survey}>
+          <Grid container>{surveySaysGrid}</Grid>
         </div>
       </div>
     </>
   );
 };
-
-const useStyles = makeStyles({ name: { SurveySays } })((theme) => ({
-  centerLock: {
-    position: "absolute",
-    top: "50%",
-    left: "calc(50% + 0px)",
-    transform: "translate(-50%, -50%)",
-  },
-  fixedSurveyCard: {
-    borderRadius: 10,
-    background: theme.palette.primary.main,
-    color: "white",
-    padding: 10,
-    height: 30,
-    position: "relative",
-    boxSizing: "border-box",
-    transition: "box-shadow 0.5s ease-in",
-  },
-  expandableSurveyCard: {
-    borderRadius: 10,
-    background: theme.palette.primary.main,
-    color: "white",
-    paddingTop: 5,
-    paddingBottom: 5,
-    paddingLeft: 10,
-    paddingRight: 10,
-    minHeight: 30,
-    position: "relative",
-    boxSizing: "border-box",
-    transition: "box-shadow 0.5s ease-in",
-  },
-  fixedSurveyCardText: {
-    whiteSpace: "nowrap",
-    overflow: "hidden",
-    textOverflow: "ellipsis",
-  },
-  expandableSurveyCardText: {
-    minHeight: 20,
-  },
-  default: {
-    boxShadow: "1px 6px 6px -3px lightblue",
-  },
-  completeSatisfied: {
-    boxShadow: "1px 6px 6px -3px lightgreen",
-  },
-  completeUnsatisfied: {
-    boxShadow: "1px 6px 6px -3px darkgrey",
-  },
-  survey: {
-    padding: theme.spacing(2),
-    maxWidth: 500,
-    marginLeft: "auto",
-    marginRight: "auto",
-  },
-  bodyRoot: {
-    width: "100%",
-    backgroundColor: "#212629",
-    display: "flex",
-    justifyContent: "center",
-    alignItems: "center",
-    padding: 10,
-    boxSizing: "border-box",
-  },
-  bodyNoMedia: {
-    // height: "60%",
-  },
-  bodyMedia: {
-    // height: "40%",
-  },
-  surveyInnerBoard: {
-    width: "100%",
-    backgroundColor: "#212629",
-    boxSizing: "border-box",
-    border: "5px solid lightblue",
-    borderRadius: 30,
-  },
-  surveyQuestion: {
-    color: "white",
-  },
-}));
 
 export default withLocation(SurveySays);
