@@ -4,7 +4,7 @@ Permission to use, copy, modify, and distribute this software and its documentat
 
 The full terms of this copyright and license should always be found in the root directory of this software deliverable as "license.txt" and if these terms are not found with this software, please contact the USC Stevens Center for the full license.
 */
-import { cyMockDefault, cyMockDialog, cySetup } from "../support/functions";
+import { cyMockDefault, cyMockDialog } from "../support/functions";
 
 describe("Guest prompt", () => {
   it("appears when no guest param", () => {
